refactor Acronyms settings into the Acronyms  file
